core: add initial code of the core sensors library
This patch is adding initial code of the core library.
It includes Android Makefile (soong and gnu make) as well Linux build
capabilities (cmake).
The main interface for the external wrappers is the ISTMSensorsHAL
class.
It introduces 'private' sensors that are part of Android but without
using any reference to the Android source code.
Data are exchanged with wrappers through the ISTMSensorsCallback class
interface.

Signed-off-by: Denis Ciocca <denis.ciocca@st.com>
Change-Id: I9e78a118e27358f9e615201810ff06a7d482a0fb