core: sensors added to the list must use sequential handle from 1
This patch is adding the test and related core code to allow to add
a new sensor to the list if and only if the handle value is sequential
to the list.
This is done to avoid the needs to check if handles are valid while
using the library, making sure that handles from 1 to list size are all
valid.

Signed-off-by: Denis Ciocca <denis.ciocca@st.com>
Change-Id: I8208076f6f67b2c36e983cf32ec855ee8ca7c877