core: fix wait-for-flush CTS issue revealed in many test cases
This patch is fixing an annoying issue with flush events while in
batch mode.
One of the issue is related to the fact that flush requested are
on a stack and we have to check all the elements on the stack
when we receive a sample. In many cases when last sample was
received after a flush more then one flush event was on the stack.
In this case, it was left there till the next batch read. And of
course it was wrong.

The second issue is related to the fact that the driver is posting the
timestamp of the flush equal to the last ever read sample. In case the
flush is requested immediately after power on, timestamp of the flush
event was basically older then the enable of the sensor. Hal was
discarding the sample during ValidateDataToPush().
We now customize the behavior of this function for SWSensors so that
we ignore the fact that sometimes we could receive old samples.

Signed-off-by: Denis Ciocca <denis.ciocca@st.com>
Change-Id: I4654a7dd2ac808a056dc694478822efd9f74da07