added stm sensors libraries support (new libraries interfaces)
This is a big change in the hal architecture, the main idea is that
too many choices are as bad as no choices.
Let's integrate mock libraries in the hal such that testing becomes
super easy and let's just replace the makefiles and the libraries
archive when needed.
Let's get help from the compiler and get rid of the ifdef statements
so that all the code is always compiled.

Signed-off-by: Denis Ciocca <denis.ciocca@st.com>
Change-Id: I23e29382a6a385e5c643bffcd4130edb2d8930c3