core: added logic to allow core to send save/load requests
This patch is adding a new function in the SensorBase such that
libraries are initialized separately from CustomInit.
This is done mainly to allow sensors to load resources during library
initialization (CustomInit is executed too early).
We now make use of callback interfaces for saving/loading accel, gyro
and magn biases.

Signed-off-by: Denis Ciocca <denis.ciocca@st.com>
Change-Id: I39cf45d12af094540159d3ade5fa165efaa3646b