core: fix onSave and onLoad data request functions
Changes related to "rework on graph class template" broke the
onLoadDataRequest() function design since it was called without the
correct updated hooks.
This patch introduce a new function postSetup that is able to load
data after the hook has been setup.
It's also fixing when the warning is printed in the classes, as
documented, function returns > 0 value in case of success.

Signed-off-by: Denis Ciocca <denis.ciocca@st.com>
Change-Id: I1b8dff61d0d8bf2fed97cec44e832654a610288f
Reviewed-on: https://sczcxd1104.scz.st.com/gerrit/c/android/stm-hal-iio/+/512
Tested-by: CI STM MSD <aosp-ger@st.com>