legacy: let's specifically enable build for the HAL and allow name override
Legacy HAL is built as a dynamic library. In order to be used as it
is, library name must be equal to sensors.variant.so (check
libhardware/hardware.c).
In order to avoid conflicts in names, now you have to add the
following to your device.mk:

TARGET_USES_STM_LEGACY_SENSORSHAL := true

If you want to build HAL with a different name (maybe because of
multi-hal), add the following to your device.mk:

STM_LEGACY_SENSORSHAL_NAME_OVERRIDE := sensors.stm.default

Signed-off-by: Denis Ciocca <denis.ciocca@st.com>
Change-Id: I949e21802b4b91516b8c99df091cf2fb9a8e98b0
Reviewed-on: https://sczcxd1104.scz.st.com/gerrit/c/android/stm-hal-iio/+/532
Tested-by: CI STM MSD <aosp-ger@st.com>