2.0: wait enough space available before write to FMQ
The Event FMQ is a synchronized FMQ, which means that
any attempt to write more events to the FMQ than the
available space allows results in a failed write.
In such case, the HAL should determine whether to write
the current set of events as two smaller groups of
events or to write all of the events together when
enough space is available.

Signed-off-by: Mario Tesi <mario.tesi@st.com>
Change-Id: I41b040bb37d4df0606fd2e40676377eb86a9007b
Reviewed-on: https://gerrit.st.com/c/linuxandroidopen/stm-hal-iio/+/303573
Reviewed-by: Matteo DAMENO <matteo.dameno@st.com>
Tested-by: CITOOLS <MDG-smet-aci-reviews@list.st.com>