core: add error message to notify inconsistency of custom IIO types
The STMicroelectronics MEMS IIO drivers include the management of custom types
and events using a dedicated stm_iio_types.h file. This file must be integrated
into the HAL and then recompiled to share the same custom types with the IIO
Drivers, including those related to Flush events. If the SensorHAL was compiled
without this dependency, the event handling could receive event types from IIO
drivers that are not recognized, causing the system to not work correctly.
To highlight this misalignment, a useful debug log message has been added to
notify the inconsistency of these custom IIO event types between Drivers and
SensorHAL layers. This allows the customer to immediately identify the issue and
solve it during the integration phase.

Signed-off-by: mariotesi <mario.tesi@st.com>
Change-Id: I694e71cdd78bae26d40bf359cf42762eb49b3745
Reviewed-on: https://gerrit.st.com/c/linuxandroidopen/stm-hal-iio/+/338603
Reviewed-by: Matteo DAMENO <matteo.dameno@st.com>
Tested-by: CITOOLS <MDG-smet-aci-reviews@list.st.com>