core, aidl, hidl, multi-hal, linux, legacy: add support for temperature-based gyro calibration library
Added support for temperature-based gyroscope calibration
algo library.

The mock version of the library in installed with this patch
in the following relative sensor hal path:
core/libs/gyro-temperature-calibration

When the real gyro-temperature-calibration installation
package is executed the mock version is replaced by the
real version.

The temperature-based gyroscope calibration state is saved
to ensure data persistence.

Signed-off-by: mariotesi <mario.tesi@st.com>
Change-Id: Ia73b06db2e5d885db32aa6e5b81e4bd34d787041