core, hidl, aidl, multi-hal, legacy, linux: add geomagnetic rotation vector support
This patch introduces support for the Android geomagnetic sensor type.
This sensor is of the composite type and is implemented by the
geomag-fusion library.

In the current version the proprietary library is replaced with
a mock version.

When the real geomagnetic proprietary library installation package is
executed, the mock version is replaced by the real version.

Signed-off-by: mariotesi <mario.tesi@st.com>
Change-Id: I14f67d91e208daeeb43b000405846112f16d436e