legacy, multi-hal: apply convertion in temperature sensor resolution
In the IIO framework the temperature is expressed in milli degrees
centigrade while in android the unit of measurement is celsius, so
in order to correctly interpret this data in Android, it is also
necessary to convert the resolution and the maxRange info.
These fixes had already been made in the AIDL and HIDL interfaces,
the porting on multi-hal and legacy was missing.

Signed-off-by: mariotesi <mario.tesi@st.com>
Change-Id: I7491d317c99ebbfac2faff83995f90d007456be2