legacy, core: fix incorrect scale for magnetometer
Magnetometers in Android have a full scale expressed
in uTesla while in Linux IIO they are expressed in Gauss,
so a conversion is necessary.
The Gauss to uT conversion function is implemented in the
static class Conversion, however importing this class into
the Convert.cpp file for the legacy HAL results in an error
in the redefinition of the Earth's gravitational constant,
so it was necessary to rename it.

Signed-off-by: mariotesi <mario.tesi@st.com>
Change-Id: I3d535f4a812a4f36c9ab88c9f6caa41e12bb188c